# api/src/main/resources/swagger.yml
openapi: 3.0.1
info:
  version: ''
  title: Link API
servers:
  - url: ''
    description: This server
security:
  - oauth: []
  - basic: [ ]
...
paths:
...
  /api/user:
    get:
      tags:
        - Users
      summary: Search all users
      operationId: searchUsers
      parameters:
        - name: includeDisabled
          in: query
          description: Whether or not to include disabled users
          schema:
            type: boolean
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        description: 'A user object created'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/user/$bulk:
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createBulkUsers
      requestBody:
        description: 'An array of user objects that should be created'
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/FHIR.BulkCreateUserOperationOutcome'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/user/{userId}:
    put:
      tags:
        - Users
      summary: Update a user
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: string
      requestBody:
        description: 'The "enabled" property of the User is ignored. It can only be set via DELETE'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Users
      summary: Delete (disable) a user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: The id of the user to delete/disable
          required: true
          schema:
            type: string
      responses:
        '204':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/user/{userId}/$change-password:
    put:
      tags:
        - Users
      summary: Change a user's password
      operationId: changeUserPassword
      parameters:
        - name: userId
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tenant:
    get:
      tags:
        - Tenants
      summary: Search all tenants
      operationId: searchTenants
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchTenantResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Tenants
      summary: Create a new tenant. If id is specified, will make sure that the tenant does not already exist and then create the tenant with the specified id
      operationId: createTenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tenant'
          'application/x-yaml':
            schema:
              $ref: '#/components/schemas/Tenant'
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tenant/{id}:
    get:
      tags:
        - Tenants
      summary: Get a single tenant
      operationId: getTenant
      parameters:
        - name: id
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Tenants
      summary: Updates an existing tenant
      operationId: updateTenant
      parameters:
        - name: id
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tenant'
          'application/x-yaml':
            schema:
              $ref: '#/components/schemas/Tenant'
      responses:
        '200':
          description: successful response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Tenants
      summary: Delete a tenant
      operationId: deleteTenant
      parameters:
        - name: id
          in: path
          description: The id of the tenant
          required: true
          schema:
            type: string
      responses:
        '204':
          description: successful response with no content returned
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    SearchTenantResponse:
      type: object
      properties:
        id:
          type: string
          description: 'A unique ID of the tenant. Cannot contain special characters other than dashes and underscores.'
        name:
          type: string
          description: 'The internal name of the tenant, such as the facility''s name.'
        retentionPeriod:
          type: string
          description: ISO 8601 formatted duration in which to keep data for each tenant.
          default: 'P3M'
        cdcOrgId:
          type: string
          description: 'Organization ID for the tenant according to CDC. Used by the Measure Reporting Plan functionality, and when bundling to submit to NHSN; an Organization resource is created uses this as an identifier of the facility.'
    Tenant:
      type: object
      required:
        - cdcOrgId
        - bundling
      properties:
        id:
          type: string
          description: 'A unique ID of the tenant. Cannot contain special characters other than dashes and underscores.'
        name:
          type: string
          description: 'The internal name of the tenant, such as the facility''s name.'
        cdcOrgId:
          type: string
          description: 'Organization ID for the tenant according to CDC. Used by the Measure Reporting Plan functionality, and when bundling to submit to NHSN; an Organization resource is created uses this as an identifier of the facility.'
        connectionString:
          type: string
          description: 'The name of the database to use for the tenant''s PHI/PII data. The same connection string for the database server is used for each tenant, but a different database on the server is used for each tenant.'
        bundling:
          allOf:
            - $ref: '#/components/schemas/Bundling'
            - description: 'Configuration properties for how to bundle the final submission, including contact information that is used when creating an Organization resource that represents the sending facility (the tenant).'
        scheduling:
          allOf:
            - $ref: '#/components/schemas/Schedule'
            - description: Schedule configuration for the tenant that can be used to automatically begin processes such as querying for patient lists, data retention, and automatic generation and submission of reports.
        events:
          allOf:
            - $ref: '#/components/schemas/Events'
            - description: 'Custom classes/functionality that can be turned on at various points during the report generation pipeline.'
        queryList:
          allOf:
            - $ref: '#/components/schemas/QueryList'
            - description: 'Configuration that should be used for /api/{tenantId}/poi/$query-list to query the EHR''s FHIR List for patients of interest'
        fhirQuery:
          allOf:
            - $ref: '#/components/schemas/FhirQuery'
            - description: 'Configuration to use when querying patients of interest for their clinical data, such as Encounter, Condition, etc. Includes authentication strategy that is needed to successfully query the EHR''s FHIR API.'
        reportingPlan:
          allOf:
            - $ref: '#/components/schemas/ReportingPlan'
            - description: 'Configuration for how to query the MRP (reporting plan) interface at CDC/NHSN to determine if a facility/tenant is signed up to report during the calculated reporting period.'
        retentionPeriod:
          type: string
          description: ISO 8601 formatted duration in which to keep data for each tenant.
          default: 'P3M'
    Bundling:
      type: object
      required:
        - name
      properties:
        bundleType:
          type: string
          enum:
            - COLLECTION
            - BATCH
          default: COLLECTION
          description: 'The type of submission bundle to create. Bundle types are defined in https://hl7.org/fhir/codesystem-bundle-type.html'
        includesCensuses:
          type: boolean
          default: true
          description: 'Whether to include censuses in the submission bundle.'
        mergeCensuses:
          type: boolean
          default: true
          description: 'Whether to merge multiple censuses into a single list.'
        promoteLineLevelResources:
          type: boolean
          default: false
          description: 'Whether to move contained line-level resources to the top level of the submission bundle.'
        npi:
          type: string
          description: 'The National Provider Identifier (NPI) of the organization/facility the system is bundling/submitting for, to be used in the submission Bundle''s Organization.identifier field'
        name:
          type: string
          description: 'The name of the organization/facility the system is bundling/submitting for, to be used in the submission Bundle''s Organization.name field'
        phone:
          type: string
          description: 'The phone of the organization/facility the system is bundling/submitting for, to be used in the submission Bundle''s Organization.name field'
        email:
          type: string
          description: 'The email of the organization/facility the system is bundling/submitting for, to be used in the submission Bundle''s Organization.name field'
        address:
          allOf:
            - $ref: '#/components/schemas/Address'
            - description: 'The address of the organization/facility the system is bundling/submitting for, to be used in the submission Bundle''s Organization.name field'
    Address:
      type: object
      properties:
        addressLine:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    Schedule:
      type: object
      description: |
        See <a href="https://productresources.collibra.com/docs/collibra/latest/Content/Cron/co_spring-cron.htm" target="_new">Spring Cron syntax</a> for more info.<br/>
        See <a href="https://crontab.guru/" target="_new">Crontab.guru</a> for more info (note: seconds aren't accounted for in this tool).
      properties:
        queryPatientListCron:
          type: string
          description: 'The CRON-formatted schedule to use querying patient list information from the EHR. The equivalent of manually calling /api/{tenantId}/poi/$query-list. Cron expression formatting can be found in https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm'
        dataRetentionCheckCron:
          type: string
          description: 'The CRON-formatted schedule to use for data retention checks (removing data that is outside the tenant''s configured retention period).'
        generateAndSubmitReports:
          type: array
          items:
            $ref: '#/components/schemas/GenerateReport'
          description: 'Configuration for when to automatically generate and submit reports.'
...
    Events:
      type: object
      properties:
        BeforeMeasureResolution:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterMeasureResolution:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        OnRegeneration:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforePatientOfInterestLookup:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterPatientOfInterestLookup:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforePatientDataQuery:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterPatientResourceQuery:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterPatientDataQuery:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterApplyConceptMaps:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforePatientDataStore:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterPatientDataStore:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforeMeasureEval:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterMeasureEval:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforeReportStore:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterReportStore:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        BeforeBundling:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
        AfterBundling:
          type: array
          items:
            $ref: '#/components/schemas/EventOptions'
...
    QueryList:
      type: object
      properties:
        fhirServerBase:
          type: string
          description: The URL to the facility FHIR endpoint
        lists:
          type: array
          items:
            $ref: '#/components/schemas/EhrPatientList'
...
    FhirQuery:
      type: object
      properties:
        fhirServerBase:
          type: string
          description: The URL to the facility FHIR endpoint
        authClass:
          type: string
          description: 'The class that should be used (if any) to authenticate queries to the EHR''s FHIR server.'
          enum:
            - com.lantanagroup.link.query.auth.BasicAuth
            - com.lantanagroup.link.query.auth.BasicAuthAndApiKeyHeader
            - com.lantanagroup.link.query.auth.AzureAuth
            - com.lantanagroup.link.query.auth.CernerAuth
            - com.lantanagroup.link.query.auth.EpicAuth
            - com.lantanagroup.link.query.auth.TokenAuth
        parallelPatients:
          type: number
          description: 'The number of patients to query for in parallel using separate threads. Should not be greater than the number of cores available to the installation.'
        basicAuth:
          allOf:
            - $ref: '#/components/schemas/BasicAuth'
            - description: 'Configuration used by BasicAuth implementation'
        basicAuthAndApiKey:
          allOf:
            - $ref: '#/components/schemas/BasicAuthAndApiKey'
            - description: 'Configuration used by basicAuthAndApiKeyHeader implementation'
        tokenAuth:
          allOf:
            - $ref: '#/components/schemas/TokenAuth'
            - description: 'Configuration used by TokenAuth implementation'
        azureAuth:
          allOf:
            - $ref: '#/components/schemas/AzureAuth'
            - description: 'Configuration used by AzureAuth implementation'
        cernerAuth:
          allOf:
            - $ref: '#/components/schemas/CernerAuth'
            - description: 'Configuration used by CernerAuth implementation'
        epicAuth:
          allOf:
            - $ref: '#/components/schemas/EpicAuth'
            - description: 'Configuration used by EpicAuth implementation'
        queryPlans:
          type: object
          properties:
            measureId:
              $ref: '#/components/schemas/QueryPlan'
    BasicAuth:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
    TokenAuth:
      type: object
      properties:
        token:
          type: string
...
    ReportingPlan:
      type: object
      required:
        - url
        - nhsnOrgId
        - planNames
      properties:
        enabled:
          type: boolean
          default: false
        url:
          type: string
        nhsnOrgId:
          type: string
        planNames:
          type: object
          required:
            - measureId
          properties:
            measureId:
              type: string
              description: Key is the  measure id registered in Link, value is the NHSN plan ID
              example: nhsnPlanId
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
...
    FHIR.BulkCreateUserOperationOutcome:
      type: object
      example: {
        "issue": [ {
          "severity": "info",
          "diagnostics": "Successfully created user with email XX@YY.com"
        }, {
          "severity": "error",
          "diagnostics": "Error creating user with email AA@BB.com: A user with that email already exists"
        } ]
      }
      properties:
        extension:
          type: array
          items:
            $ref: '#/components/schemas/FHIR.Extension'
        issue:
          type: array
          items:
            $ref: '#/components/schemas/FHIR.OperationOutcomeIssue'
...
    FHIR.OperationOutcomeIssue:
      type: object
      properties:
        severity:
          type: string
        code:
          type: string
        diagnostics:
          type: string
        location:
          type: array
          items:
            type: string
    FHIR.Extension:
      type: object
      properties:
        url:
          type: string
        valueInteger:
          type: number
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    oauth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ''
          tokenUrl: ''
          scopes: { }
    basic:
      type: http
      scheme: basic
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'